fix: calling FederatedEventTarget.off() with original listener now works when using signal
The initial implementation for signal support wrapped the provided listener
function to add an additional check for the signal's state. This made it
impossible to manually remove the listener by calling off() as there was no way
to get a reference to the wrapped listener function. Luckily, this
additional check was superfluous given the single threaded nature of
JavaScript's event loop.